refatco : map implementation
